moved shared dialog components to DialogComponents
